Fix lint + remove redundant permission requests
