+++ app/src/main/java/dev/arkbuilders/arkretouch/utils/permission/PermissionsHelper.kt
@@ -3,8 +3,12 @@ package dev.arkbuilders.arkretouch.utils.permission
 import androidx.activity.result.ActivityResultLauncher
 import androidx.activity.result.contract.ActivityResultContract
 import androidx.activity.result.contract.ActivityResultContracts
+import androidx.core.content.ContextCompat
 import android.Manifest
+import android.content.Context
+import android.content.pm.PackageManager
 import android.os.Build
+import android.os.Environment
 import dev.arkbuilders.arkretouch.BuildConfig
 
 object PermissionsHelper {
@@ -22,6 +26,15 @@ object PermissionsHelper {
             launcher.launch(packageUri)
         } else {
             launcher.launch(Manifest.permission.WRITE_EXTERNAL_STORAGE)
+        }
+    }
+
+    fun Context.isWritePermissionGranted(): Boolean {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
+            Environment.isExternalStorageManager()
+        } else {
+            val checkSelfPermission = ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE)
+            checkSelfPermission == PackageManager.PERMISSION_GRANTED
         }
     }
 }
